# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/statsbomb/sync": {
      "post": {
        "description": "Fetches data from StatsBomb Open Data repository and syncs it with the local database",
        "operationId": "StatsBombController_syncData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Data synchronization completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "StatsBomb data synchronization completed successfully"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-15T10:30:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error during synchronization"
          }
        },
        "summary": "Sync StatsBomb data to database",
        "tags": [
          "StatsBomb Adapter"
        ]
      }
    },
...
    "/statsbomb/test/sync-single-match/{matchId}": {
      "post": {
        "description": "Tests syncing teams, players, and all events (goals, cards, substitutions) from a specific match ID to debug the process",
        "operationId": "StatsBombController_testSyncSingleMatch",
        "parameters": [
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Single match sync test result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "matchId": {
                      "type": "number"
                    },
                    "playersCreated": {
                      "type": "number"
                    },
                    "eventsCreated": {
                      "type": "number"
                    },
                    "details": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Test sync complete match data from a single match",
        "tags": [
          "StatsBomb Adapter"
        ]
      }
    },
...
  },
  "info": {
    "title": "I Watch Football API",
    "description": "The API Docs for I Watch Football",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}